Allow BufWriter to buffer multiple packets
Normally, after writing a full packet we just flush the BufWriter. This
is because without flushing, a ping might end up waiting for a long time
if there is no bulk traffic on the socket.

With this change, after writing a packet, instead of flushing
immediately, we peek at the packet queue to see if there is a data
packet available right now. If there is, we don't flush and we write all
the available packets into the BufWriter until we run out.

This gave us around 25% speed increase in downloads. Massive low-hanging
fruit.